Sua co che nhan su kien cua cart

+++ backend/services/cart-service/cart/src/main/java/vn/uit/edu/msshop/cart/adapter/in/web/CartProductEventListener.java
@@ -1,5 +1,6 @@
 package vn.uit.edu.msshop.cart.adapter.in.web;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
@@ -10,6 +11,8 @@ import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
 import vn.uit.edu.msshop.cart.adapter.in.web.mapper.CartWebMapper;
+import vn.uit.edu.msshop.cart.adapter.out.event.EventDocument;
+import vn.uit.edu.msshop.cart.adapter.out.event.EventDocumentRepository;
 import vn.uit.edu.msshop.cart.application.dto.command.DeleteCartItemCommand;
 import vn.uit.edu.msshop.cart.application.dto.command.UpdateCartInfoCommand;
 import vn.uit.edu.msshop.cart.application.port.in.DeleteCartItemUseCase;
@@ -28,17 +31,25 @@ public class CartProductEventListener {
     private final UpdateCartInfoUseCase updateCartInfoUseCase;
     private final CartWebMapper mapper;
     private final DeleteCartItemUseCase deleteItemUseCase;
+    private final EventDocumentRepository eventDocumentRepo;
+
     @KafkaHandler
     public void updateCartItem(ProductUpdated event) {
-        Set<String> userIds = variantToUserPort.getByVariantId(new VariantId(event.getVariantId()));
-        List<UpdateCartInfoCommand> commands = userIds.stream().map(item->mapper.toCommand(event, item)).toList();
-        updateCartInfoUseCase.updateInfo(commands);
+        if(!eventDocumentRepo.existsById(event.getEventId())) {
+            Set<String> userIds = variantToUserPort.getByVariantId(new VariantId(event.getVariantId()));
+            List<UpdateCartInfoCommand> commands = userIds.stream().map(item->mapper.toCommand(event, item)).toList();
+            updateCartInfoUseCase.updateInfo(commands);
+            eventDocumentRepo.save(EventDocument.builder().eventId(event.getEventId()).receiveAt(Instant.now()).build());
+        }
     }
     @KafkaHandler
     public void onVariantDelete(ProductDeleted event) {
-        Set<String> userIds = variantToUserPort.getByVariantId(new VariantId(event.getVariantId()));
-        List<DeleteCartItemCommand> commands = userIds.stream().map(item->new DeleteCartItemCommand(new UserId(UUID.fromString(item)), new VariantId(event.getVariantId()))).toList();
-        deleteItemUseCase.deleteManyItems(commands);
+        if(!eventDocumentRepo.existsById(event.getEventId())) {
+            Set<String> userIds = variantToUserPort.getByVariantId(new VariantId(event.getVariantId()));
+            List<DeleteCartItemCommand> commands = userIds.stream().map(item->new DeleteCartItemCommand(new UserId(UUID.fromString(item)), new VariantId(event.getVariantId()))).toList();
+            deleteItemUseCase.deleteManyItems(commands);
+            eventDocumentRepo.save(EventDocument.builder().eventId(event.getEventId()).receiveAt(Instant.now()).build());
+        }
 
     }
 }
